Feature  ore prices based on materials

+++ Code/BuyOrderCalulator.Domain/OrderItem.cs
@@ -4,7 +4,7 @@ namespace BuyOrderCalc.Domain
     public class OrderItem: Entity
     {
         public int Quantity { get; set; }
-        public int FixedUnitPrice { get; set; }
+        public double FixedUnitPrice { get; set; }
         public double FixedCorpCreditPercent { get; set; }
 
         public int ItemId { get; set; }
